Updated SushiYieldSource to use the new Yield Source Interface.
The interface has been changed because it conflicted with the ERC20
interface.  Given the likelyhood of the yield source being a token, we
had to make this breaking change.

Note that the change is not yet live on mainnet; so the tests currently
fail.

There should be unit tests, either way, that can function without
mainnet.